Remove web worker + dep upgrades
