+++ apps/www/package.json
@@ -18,7 +18,6 @@
 		"subset-serif-font": "pnpm exec glyphhanger --formats=woff2 --whitelist=Dicephrase --subset=./public/fonts/DM_Serif_Display/DMSerifDisplay-Regular.ttf"
 	},
 	"dependencies": {
-		"@ct/tworker": "workspace:^0.0.1",
 		"gen-utils": "workspace:0.0.4",
 		"solid-js": "1.6.9"
 	},
@@ -28,18 +27,18 @@
 		"@types/escodegen": "^0.0.7",
 		"@types/html-minifier-terser": "^7.0.0",
 		"@types/node": "^18.11.18",
-		"@typescript-eslint/eslint-plugin": "^5.48.1",
-		"@typescript-eslint/parser": "^5.48.1",
-		"@typescript-eslint/typescript-estree": "^5.48.1",
-		"@vanilla-extract/css": "^1.9.2",
-		"@vanilla-extract/vite-plugin": "^3.7.0",
+		"@typescript-eslint/eslint-plugin": "^5.49.0",
+		"@typescript-eslint/parser": "^5.49.0",
+		"@typescript-eslint/typescript-estree": "^5.49.0",
+		"@vanilla-extract/css": "^1.9.3",
+		"@vanilla-extract/vite-plugin": "^3.7.1",
 		"astray": "^1.1.1",
 		"esbuild": "^0.15.18",
 		"escodegen": "^2.0.0",
-		"eslint": "^8.31.0",
+		"eslint": "^8.32.0",
 		"eslint-config-prettier": "^8.6.0",
 		"eslint-import-resolver-typescript": "^3.5.3",
-		"eslint-plugin-import": "^2.27.4",
+		"eslint-plugin-import": "^2.27.5",
 		"eslint-plugin-jsx-a11y": "^6.7.1",
 		"glyphhanger": "^5.0.0",
 		"html-minifier-terser": "^7.1.0",
